#545 - refactor outdated signer code
