+++ package.json
@@ -59,21 +59,22 @@
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@utexo/rgb-sdk-core": "file:../rgb-sdk-core",
     "@bitcoindevkit/bdk-wallet-node": "^0.2.0",
     "@bitcoindevkit/bdk-wallet-web": "^0.2.0",
     "@noble/hashes": "^2.0.1",
+    "@scure/btc-signer": "^2.0.1",
+    "@utexo/rgb-lib": "0.3.0-beta.13",
+    "@utexo/rgb-sdk-core": "file:../rgb-sdk-core",
     "axios": "^1.8.4",
     "bare-node-runtime": "^1.1.4",
     "bitcoinjs-lib": "^6.1.7",
     "form-data": "^4.0.0",
-    "@utexo/rgb-lib": "0.3.0-beta.13",
     "ripemd160": "^2.0.3"
   },
   "optionalDependencies": {
-    "@utexo/rgb-lib-linux-x64": "0.3.0-beta.13",
+    "@utexo/rgb-lib-darwin-arm64": "0.3.0-beta.13",
     "@utexo/rgb-lib-linux-arm64": "0.3.0-beta.13",
-    "@utexo/rgb-lib-darwin-arm64": "0.3.0-beta.13"
+    "@utexo/rgb-lib-linux-x64": "0.3.0-beta.13"
   },
   "keywords": [
     "rgb",
