--- src/binding/index.ts
@@ -1,7 +1,0 @@
-export {
-  NodeRgbLibBinding,
-  generateKeys,
-  restoreWallet,
-  restoreFromVss,
-  type RgbLibGeneratedKeys,
-} from './NodeRgbLibBinding';
