--- src/constants.ts
@@ -1,14 +1,0 @@
-/**
- * Shared constants for the RGB SDK
- *
- * This file re-exports constants from the organized constants directory
- * for backward compatibility and convenience.
- *
- * @deprecated Import from './constants' directly or use domain-specific exports
- * @see ./constants/derivation
- * @see ./constants/network
- * @see ./constants/defaults
- */
-
-// Re-export all constants from organized structure
-export * from './constants/index';
