+++ src/types/utexo.ts
@@ -1,122 +1,122 @@
 import { InvoiceRequest, SendResult } from '@utexo/rgb-sdk-core';
 
-/**
- * UTEXO Protocol Types
- */
+// /**
+//  * UTEXO Protocol Types
+//  */
 
-export type PublicKeys = {
-  xpub: string;
-  accountXpubVanilla: string;
-  accountXpubColored: string;
-  masterFingerprint: string;
-};
+// export type PublicKeys = {
+//   xpub: string;
+//   accountXpubVanilla: string;
+//   accountXpubColored: string;
+//   masterFingerprint: string;
+// };
 
-/**
- * Lightning API Types
- */
-export interface LightningAsset {
-  /**
-   * @type {string}
-   * @memberof LightningAsset
-   */
-  assetId: string;
+// /**
+//  * Lightning API Types
+//  */
+// export interface LightningAsset {
+//   /**
+//    * @type {string}
+//    * @memberof LightningAsset
+//    */
+//   assetId: string;
 
-  /**
-   * @type {number}
-   * @memberof LightningAsset
-   */
-  amount: number;
-}
-/**
- * Request model for creating Lightning invoice.
- *
- * @export
- * @interface CreateLightningInvoiceRequestModel
- */
-export interface CreateLightningInvoiceRequestModel {
-  /**
-   * @type {number}
-   * @memberof CreateLightningInvoiceRequestModel
-   */
-  amountSats?: number;
+//   /**
+//    * @type {number}
+//    * @memberof LightningAsset
+//    */
+//   amount: number;
+// }
+// /**
+//  * Request model for creating Lightning invoice.
+//  *
+//  * @export
+//  * @interface CreateLightningInvoiceRequestModel
+//  */
+// export interface CreateLightningInvoiceRequestModel {
+//   /**
+//    * @type {number}
+//    * @memberof CreateLightningInvoiceRequestModel
+//    */
+//   amountSats?: number;
 
-  /**
-   * @type {LightningAsset}
-   * @memberof CreateLightningInvoiceRequestModel
-   */
-  asset: LightningAsset;
+//   /**
+//    * @type {LightningAsset}
+//    * @memberof CreateLightningInvoiceRequestModel
+//    */
+//   asset: LightningAsset;
 
-  /**
-   * @type {number}
-   * @memberof CreateLightningInvoiceRequestModel
-   */
-  expirySeconds?: number;
-}
+//   /**
+//    * @type {number}
+//    * @memberof CreateLightningInvoiceRequestModel
+//    */
+//   expirySeconds?: number;
+// }
 
-export interface LightningReceiveRequest {
-  lnInvoice: string;
-  expiresAt?: number;
-}
+// export interface LightningReceiveRequest {
+//   lnInvoice: string;
+//   expiresAt?: number;
+// }
 
-export interface LightningSendRequest extends SendResult {}
+// export interface LightningSendRequest extends SendResult {}
 
-export interface GetLightningSendFeeEstimateRequestModel {
-  invoice: string;
-  assetId?: string;
-}
+// export interface GetLightningSendFeeEstimateRequestModel {
+//   invoice: string;
+//   assetId?: string;
+// }
 
-export interface PayLightningInvoiceRequestModel {
-  lnInvoice: string;
-  amount?: number;
-  assetId?: string;
-  maxFee?: number;
-}
+// export interface PayLightningInvoiceRequestModel {
+//   lnInvoice: string;
+//   amount?: number;
+//   assetId?: string;
+//   maxFee?: number;
+// }
 
-export interface PayLightningInvoiceEndRequestModel {
-  signedPsbt: string;
-  lnInvoice: string;
-}
+// export interface PayLightningInvoiceEndRequestModel {
+//   signedPsbt: string;
+//   lnInvoice: string;
+// }
 
-/**
- * Onchain API Types
- */
+// /**
+//  * Onchain API Types
+//  */
 
-export interface OnchainReceiveRequestModel extends InvoiceRequest {
-  amount: number;
-  assetId: string;
-}
+// export interface OnchainReceiveRequestModel extends InvoiceRequest {
+//   amount: number;
+//   assetId: string;
+// }
 
-export interface OnchainReceiveResponse {
-  /** Mainnet invoice */
-  invoice: string;
-}
+// export interface OnchainReceiveResponse {
+//   /** Mainnet invoice */
+//   invoice: string;
+// }
 
-export interface OnchainSendRequestModel {
-  /** Mainnet invoice */
-  invoice: string;
-  assetId?: string;
-  amount?: number;
-}
+// export interface OnchainSendRequestModel {
+//   /** Mainnet invoice */
+//   invoice: string;
+//   assetId?: string;
+//   amount?: number;
+// }
 
-export interface OnchainSendEndRequestModel {
-  /** Mainnet invoice */
-  invoice: string;
-  signedPsbt: string;
-}
+// export interface OnchainSendEndRequestModel {
+//   /** Mainnet invoice */
+//   invoice: string;
+//   signedPsbt: string;
+// }
 
-export interface OnchainSendResponse extends SendResult {}
+// export interface OnchainSendResponse extends SendResult {}
 
-export interface GetOnchainSendResponse {
-  sendId: string;
-  txid?: string;
-  status: string;
-  amount: number;
-  assetId?: string;
-  fee?: number;
-  createdAt: number;
-  completedAt?: number;
-}
+// export interface GetOnchainSendResponse {
+//   sendId: string;
+//   txid?: string;
+//   status: string;
+//   amount: number;
+//   assetId?: string;
+//   fee?: number;
+//   createdAt: number;
+//   completedAt?: number;
+// }
 
-export interface ListLightningPaymentsResponse {
-  payments: LightningSendRequest[];
-}
+// export interface ListLightningPaymentsResponse {
+//   payments: LightningSendRequest[];
+// }
