--- src/utexo/bridge/index.ts
@@ -1,16 +1,0 @@
-export { getBridgeAPI, encodeTransferStatus, TransferStatuses } from '@utexo/rgb-sdk-core';
-export type {
-  NetworkAddress,
-  TransferType,
-  Estimation,
-  BridgeInSignatureRequest,
-  BridgeInSignatureResponse,
-  SubmitTransactionRequest,
-  SubmitTransactionResponse,
-  VerifyBridgeInRequest,
-  ReceiverInvoiceResponse,
-  TokenInfo,
-  TransactionHash,
-  TransferByMainnetInvoiceResponse,
-  ApiError,
-} from '@utexo/rgb-sdk-core';
