--- src/utexo/config/index.ts
@@ -1,8 +1,0 @@
-export {
-  DEFAULT_GATEWAY_BASE_URLS,
-  DEFAULT_VSS_SERVER_URL,
-  getVssConfigs,
-  testnetPreset,
-  mainnetPreset,
-} from '@utexo/rgb-sdk-core';
-export type { ConfigOptions } from '@utexo/rgb-sdk-core';
