--- src/utexo/utils/index.ts
@@ -1,15 +1,0 @@
-export {
-  getUtxoNetworkConfig,
-  utexoNetworkMap,
-  utexoNetworkIdMap,
-  getDestinationAsset,
-  decodeBridgeInvoice,
-} from '@utexo/rgb-sdk-core';
-export type {
-  UtxoNetworkPreset,
-  UtxoNetworkMap,
-  UtxoNetworkIdMap,
-  UtxoNetworkPresetConfig,
-  NetworkAsset,
-  UtxoNetworkId,
-} from '@utexo/rgb-sdk-core';
