+++ src/utils/environment.ts
@@ -5,17 +5,3 @@ export function isNode(): boolean {
     process.versions.node != null
   );
 }
-
-export function isBrowser(): boolean {
-  return (
-    typeof window !== 'undefined' && typeof window.document !== 'undefined'
-  );
-}
-
-export type Environment = 'node' | 'browser' | 'unknown';
-
-export function getEnvironment(): Environment {
-  if (isNode()) return 'node';
-  if (isBrowser()) return 'browser';
-  return 'unknown';
-}
