--- src/wallet/index.ts
@@ -1,15 +1,0 @@
-/**
- * Wallet module exports
- *
- * This module provides the WalletManager class and related functionality for
- * managing RGB wallets, combining RGB Node API client and cryptographic operations.
- */
-
-export {
-  WalletManager,
-  createWalletManager,
-  wallet,
-  createWallet,
-  restoreFromBackup,
-} from './wallet-manager';
-export type { WalletInitParams } from './wallet-manager';
